soroban_sdk
Module auth

Auth contains types for building custom account contracts.

Structs

ContractContext

Authorization context of a single contract call.

CreateContractHostFnContext

Authorization context for create_contract host function that creates a new contract on behalf of authorizer address.

SubContractInvocation

Value of contract node in InvokerContractAuthEntry tree.

Enums

Context

Context of a single authorized call performed by an address.

ContractExecutable

Contract executable used for creating a new contract and used in CreateContractHostFnContext.

InvokerContractAuthEntry

A node in the tree of authorizations performed on behalf of the current contract as invoker of the contracts deeper in the call stack.

Traits

CustomAccountInterface

Custom account interface that a contract implements to support being used as a custom account for auth.