soroban_sdk::auth
Trait CustomAccountInterface

pub trait CustomAccountInterface {
    type Signature;
    type Error: Into<Error>;

    // Required method
    fn __check_auth(
        env: Env,
        signature_payload: Hash<32>,
        signatures: Self::Signature,
        auth_contexts: Vec<Context>,
    ) -> Result<(), Self::Error>;
}

Custom account interface that a contract implements to support being used as a custom account for auth.

Once a contract implements the interface, call to Address::require_auth for the contract’s address will call its __check_auth implementation.

Required Associated Types

type Signature

type Error: Into<Error>

Required Methods

fn __check_auth( env: Env, signature_payload: Hash<32>, signatures: Self::Signature, auth_contexts: Vec<Context>, ) -> Result<(), Self::Error>

Check that the signatures and auth contexts are valid.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.