soroban_sdk::crypto

Module bls12_381

Structs

Bls12_381

Bls12_381 provides access to curve and field arithmetics on the BLS12-381 curve.

Fp

Fp represents an element of the base field Fq of the BLS12-381 elliptic curve

Fp2

Fp2 represents an element of the quadratic extension field Fq2 of the BLS12-381 elliptic curve

Fr

Fr represents an element in the BLS12-381 scalar field, which is a prime field of order r (the order of the G1 and G2 groups). The struct is internally represented with an U256, all arithmetic operations follow modulo r.

G1Affine

G1Affine is a point in the G1 group (subgroup defined over the base field Fq) of the BLS12-381 elliptic curve

G2Affine

G2Affine is a point in the G2 group (subgroup defined over the quadratic extension field Fq2) of the BLS12-381 elliptic curve

Constants

FP2_SERIALIZED_SIZE

FP_SERIALIZED_SIZE

G1_SERIALIZED_SIZE

G2_SERIALIZED_SIZE