soroban_sdk::crypto

Struct Crypto

pub struct Crypto { /* private fields */ }

Crypto provides access to cryptographic functions.

Implementations

impl Crypto

pub fn env(&self) -> &Env

pub fn sha256(&self, data: &Bytes) -> Hash<32>

Returns the SHA-256 hash of the data.

pub fn keccak256(&self, data: &Bytes) -> Hash<32>

Returns the Keccak-256 hash of the data.

pub fn ed25519_verify( &self, public_key: &BytesN<32>, message: &Bytes, signature: &BytesN<64>, )

Verifies an ed25519 signature.

The signature is verified as a valid signature of the message by the ed25519 public key.

Panics

If the signature verification fails.

pub fn secp256k1_recover( &self, message_digest: &Hash<32>, signature: &BytesN<64>, recorvery_id: u32, ) -> BytesN<65>

Recovers the ECDSA secp256k1 public key.

The public key returned is the SEC-1-encoded ECDSA secp256k1 public key that produced the 64-byte signature over a given 32-byte message digest, for a given recovery_id byte.

pub fn secp256r1_verify( &self, public_key: &BytesN<65>, message_digest: &Hash<32>, signature: &BytesN<64>, )

Verifies the ECDSA secp256r1 signature.

The SEC-1-encoded public key is provided along with the message, verifies the 64-byte signature.