soroban_sdk::deploy
Struct Deployer

pub struct Deployer { /* private fields */ }

Deployer provides access to deploying contracts.

Implementations

impl Deployer

pub fn env(&self) -> &Env

pub fn with_current_contract( &self, salt: impl IntoVal<Env, BytesN<32>>, ) -> DeployerWithAddress

Get a deployer that deploys contract that derive the contract IDs from the current contract and provided salt.

pub fn with_address( &self, address: Address, salt: impl IntoVal<Env, BytesN<32>>, ) -> DeployerWithAddress

Get a deployer that deploys contracts that derive the contract ID from the provided address and salt.

The deployer address must authorize all the deployments.

pub fn with_stellar_asset( &self, serialized_asset: impl IntoVal<Env, Bytes>, ) -> DeployerWithAsset

Get a deployer that deploys an instance of Stellar Asset Contract corresponding to the provided serialized asset.

serialized_asset is the Stellar Asset XDR serialized to bytes. Refer to [soroban_sdk::xdr::Asset]

pub fn upload_contract_wasm( &self, contract_wasm: impl IntoVal<Env, Bytes>, ) -> BytesN<32>

Upload the contract Wasm code to the network.

Returns the hash of the uploaded Wasm that can be then used for the contract deployment.

Examples

use soroban_sdk::{BytesN, Env};

const WASM: &[u8] = include_bytes!("../doctest_fixtures/contract.wasm");

#[test]
fn test() {
    let env = Env::default();
    env.deployer().upload_contract_wasm(WASM);
}

pub fn update_current_contract_wasm( &self, wasm_hash: impl IntoVal<Env, BytesN<32>>, )

Replaces the executable of the current contract with the provided Wasm.

The Wasm blob identified by the wasm_hash has to be already present in the ledger (uploaded via [Deployer::upload_contract_wasm]).

The function won’t do anything immediately. The contract executable will only be updated after the invocation has successfully finished.

pub fn extend_ttl( &self, contract_address: Address, threshold: u32, extend_to: u32, )

Extend the TTL of the contract instance and code.

Extends the TTL of the instance and code only if the TTL for the provided contract is below threshold ledgers. The TTL will then become extend_to. Note that the threshold check and TTL extensions are done for both the contract code and contract instance, so it’s possible that one is bumped but not the other depending on what the current TTL’s are.

The TTL is the number of ledgers between the current ledger and the final ledger the data can still be accessed.

pub fn extend_ttl_for_contract_instance( &self, contract_address: Address, threshold: u32, extend_to: u32, )

Extend the TTL of the contract instance.

Same as extend_ttl but only for contract instance.

pub fn extend_ttl_for_code( &self, contract_address: Address, threshold: u32, extend_to: u32, )

Extend the TTL of the contract code.

Same as extend_ttl but only for contract code.