soroban_sdk

Macro contractfile

contractfile!() { /* proc-macro */ }

Import a contract from its WASM file, generating a constant holding the contract file.

Note that contractimport also automatically imports the contract file into a constant, and so it is usually unnecessary to use contractfile directly, unless you specifically want to only load the contract file without generating a client for it.