soroban_sdk::prng
Struct Prng

pub struct Prng { /* private fields */ }

Prng is a pseudo-random generator.

Warning

The PRNG is unsuitable for generating secrets or use in applications with low risk tolerance, see the module-level comment.

Implementations

impl Prng

pub fn env(&self) -> &Env

pub fn seed(&self, seed: Bytes)

Reseeds the PRNG with the provided value.

Warning

The PRNG is unsuitable for generating secrets or use in applications with low risk tolerance, see the module-level comment.

pub fn fill<T>(&self, v: &mut T)
where T: Fill + ?Sized,

Fills the type with a random value.

Warning

The PRNG is unsuitable for generating secrets or use in applications with low risk tolerance, see the module-level comment.

Examples

u64

let mut value: u64 = 0;
env.prng().fill(&mut value);
assert_eq!(value, 8478755077819529274);

[u8]

let mut value = [0u8; 32];
env.prng().fill(&mut value);
assert_eq!(
  value,
  [
    58, 248, 248, 38, 210, 150, 170, 117, 122, 110, 9, 101, 244, 57,
    221, 102, 164, 48, 43, 104, 222, 229, 242, 29, 25, 148, 88, 204,
    130, 148, 2, 66
  ],
);

pub fn gen<T>(&self) -> T
where T: Gen,

Returns a random value of the given type.

Warning

The PRNG is unsuitable for generating secrets or use in applications with low risk tolerance, see the module-level comment.

Examples

u64

let value: u64 = env.prng().gen();
assert_eq!(value, 8478755077819529274);

[u8; N]

let value: [u8; 32] = env.prng().gen();
assert_eq!(
  value,
  [
    58, 248, 248, 38, 210, 150, 170, 117, 122, 110, 9, 101, 244, 57,
    221, 102, 164, 48, 43, 104, 222, 229, 242, 29, 25, 148, 88, 204,
    130, 148, 2, 66
  ],
);

pub fn gen_len<T>(&self, len: T::Len) -> T
where T: GenLen,

Returns a random value of the given type with the given length.

Panics

If the length is greater than u32::MAX.

Warning

The PRNG is unsuitable for generating secrets or use in applications with low risk tolerance, see the module-level comment.

Examples

Bytes

// Get a value of length 32 bytes.
let value: Bytes = env.prng().gen_len(32);
assert_eq!(value, Bytes::from_slice(
  &env,
  &[
    58, 248, 248, 38, 210, 150, 170, 117, 122, 110, 9, 101, 244, 57,
    221, 102, 164, 48, 43, 104, 222, 229, 242, 29, 25, 148, 88, 204,
    130, 148, 2, 66
  ],
));

pub fn gen_range<T>(&self, r: impl RangeBounds<T::RangeBound>) -> T
where T: GenRange,

Returns a random value of the given type in the range specified.

Panics

If the start of the range is greater than the end.

Warning

The PRNG is unsuitable for generating secrets or use in applications with low risk tolerance, see the module-level comment.

Examples

u64

// Get a value in the range of 1 to 100, inclusive.
let value: u64 = env.prng().gen_range(1..=100);
assert_eq!(value, 46);

pub fn u64_in_range(&self, r: impl RangeBounds<u64>) -> u64

👎Deprecated: use env.prng().gen_range(…)

Returns a random u64 in the range specified.

Panics

If the range is empty.

Warning

The PRNG is unsuitable for generating secrets or use in applications with low risk tolerance, see the module-level comment.

Examples

// Get a value in the range of 1 to 100, inclusive.
let value = env.prng().u64_in_range(1..=100);
assert_eq!(value, 46);

pub fn shuffle<T>(&self, v: &mut T)
where T: Shuffle,

Shuffles a value using the Fisher-Yates algorithm.

Warning

The PRNG is unsuitable for generating secrets or use in applications with low risk tolerance, see the module-level comment.