soroban_sdk::prng
Trait Gen

pub trait Gen {
    // Required method
    fn gen(prng: &Prng) -> Self;
}

Implemented by types that support being generated by a Prng.

Required Methods

fn gen(prng: &Prng) -> Self

Generates a value of the implementing type with the Prng.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.