soroban_sdk
Struct Bytes

pub struct Bytes { /* private fields */ }

Bytes is a contiguous growable array type containing u8s.

The array is stored in the Host and available to the Guest through the functions defined on Bytes.

Bytes values can be stored as Storage, or in other types like Vec, Map, etc.

Examples

Bytes values can be created from slices:

use soroban_sdk::{Bytes, Env};

let env = Env::default();
let bytes = Bytes::from_slice(&env, &[1; 32]);
assert_eq!(bytes.len(), 32);
let mut slice = [0u8; 32];
bytes.copy_into_slice(&mut slice);
assert_eq!(slice, [1u8; 32]);

Implementations

impl Bytes

pub fn env(&self) -> &Env

pub fn as_val(&self) -> &Val

pub fn to_val(&self) -> Val

pub fn as_object(&self) -> &BytesObject

pub fn to_object(&self) -> BytesObject

impl Bytes

pub fn new(env: &Env) -> Bytes

Create an empty Bytes.

pub fn from_array<const N: usize>(env: &Env, items: &[u8; N]) -> Bytes

Create a Bytes from the array.

pub fn from_slice(env: &Env, items: &[u8]) -> Bytes

Create a Bytes from the slice.

pub fn set(&mut self, i: u32, v: u8)

Sets the byte at the position with new value.

Panics

If the position is out-of-bounds.

pub fn get(&self, i: u32) -> Option<u8>

Returns the byte at the position or None if out-of-bounds.

pub fn get_unchecked(&self, i: u32) -> u8

Returns the byte at the position.

Panics

If the position is out-of-bounds.

pub fn is_empty(&self) -> bool

Returns true if the Bytes is empty and has a length of zero.

pub fn len(&self) -> u32

Returns the number of bytes are in the Bytes.

pub fn first(&self) -> Option<u8>

Returns the first byte or None if empty.

pub fn first_unchecked(&self) -> u8

Returns the first byte.

Panics

If the Bytes is empty.

pub fn last(&self) -> Option<u8>

Returns the last byte or None if empty.

pub fn last_unchecked(&self) -> u8

Returns the last byte.

Panics

If the Bytes is empty.

pub fn remove(&mut self, i: u32) -> Option<()>

Removes the byte at the position.

Returns None if out-of-bounds.

pub fn remove_unchecked(&mut self, i: u32)

Removes the byte at the position.

Panics

If the position is out-of-bounds.

pub fn push_back(&mut self, x: u8)

Adds the byte to the back.

Increases the length by one and puts the byte in the last position.

pub fn pop_back(&mut self) -> Option<u8>

Removes and returns the last byte or None if empty.

pub fn pop_back_unchecked(&mut self) -> u8

Removes and returns the last byte or None if empty.

Panics

If the Bytes is empty.

pub fn insert(&mut self, i: u32, b: u8)

Insert the byte at the position.

Panics

If the position is out-of-bounds.

pub fn insert_from_bytes(&mut self, i: u32, bytes: Bytes)

Insert the bytes at the position.

Panics

If the position is out-of-bounds.

pub fn insert_from_array<const N: usize>(&mut self, i: u32, array: &[u8; N])

Insert the bytes at the position.

Panics

If the position is out-of-bounds.

pub fn insert_from_slice(&mut self, i: u32, slice: &[u8])

Insert the bytes at the position.

Panics

If the position is out-of-bounds.

pub fn append(&mut self, other: &Bytes)

Append the bytes.

pub fn extend_from_array<const N: usize>(&mut self, array: &[u8; N])

Extend with the bytes in the array.

pub fn extend_from_slice(&mut self, slice: &[u8])

Extend with the bytes in the slice.

pub fn copy_from_slice(&mut self, i: u32, slice: &[u8])

Copy the bytes from slice.

The full number of bytes in slice are always copied and Bytes is grown if necessary.

pub fn copy_into_slice(&self, slice: &mut [u8])

Copy the bytes into the given slice.

Panics

If the output slice and bytes are of different lengths.

pub fn slice(&self, r: impl RangeBounds<u32>) -> Self

Returns a subset of the bytes as defined by the start and end bounds of the range.

Panics

If the range is out-of-bounds.

pub fn iter(&self) -> BytesIter

pub fn to_buffer<const B: usize>(&self) -> BytesBuffer<B>

Copy the bytes into a buffer of given size.

Returns the buffer and a range of where the bytes live in the given buffer.

Suitable when the size of the bytes isn’t a fixed size but it is known to be under a certain size, or failure due to overflow is acceptable.

Panics

If the size of the bytes is larger than the size of the buffer. To avoid this, first slice the bytes into a smaller size then convert to a buffer.