soroban_sdk

Struct BytesN

pub struct BytesN<const N: usize>(/* private fields */);

BytesN is a contiguous fixed-size array type containing u8s.

The array is stored in the Host and available to the Guest through the functions defined on Bytes.

Bytes values can be stored as Storage, or in other types like Vec, Map, etc.

Examples

BytesN values can be created from arrays:

use soroban_sdk::{Bytes, BytesN, Env};

let env = Env::default();
let bytes = BytesN::from_array(&env, &[0; 32]);
assert_eq!(bytes.len(), 32);

BytesN and Bytes values are convertible:

use soroban_sdk::{Bytes, BytesN, Env};

let env = Env::default();
let bytes = Bytes::from_slice(&env, &[0; 32]);
let bytes: BytesN<32> = bytes.try_into().expect("bytes to have length 32");
assert_eq!(bytes.len(), 32);

Implementations

impl<const N: usize> BytesN<N>

pub fn env(&self) -> &Env

pub fn as_val(&self) -> &Val

pub fn to_val(&self) -> Val

pub fn as_object(&self) -> &BytesObject

pub fn to_object(&self) -> BytesObject

pub fn from_array(env: &Env, items: &[u8; N]) -> BytesN<N>

Create a BytesN from the slice.

pub fn set(&mut self, i: u32, v: u8)

Sets the byte at the position with new value.

Panics

If the position is out-of-bounds.

pub fn get(&self, i: u32) -> Option<u8>

Returns the byte at the position or None if out-of-bounds.

pub fn get_unchecked(&self, i: u32) -> u8

Returns the byte at the position.

Panics

If the position is out-of-bounds.

pub fn is_empty(&self) -> bool

Returns true if the Bytes is empty and has a length of zero.

pub fn len(&self) -> u32

Returns the number of bytes are in the Bytes.

pub fn first(&self) -> Option<u8>

Returns the first byte or None if empty.

pub fn first_unchecked(&self) -> u8

Returns the first byte.

Panics

If the Bytes is empty.

pub fn last(&self) -> Option<u8>

Returns the last byte or None if empty.

pub fn last_unchecked(&self) -> u8

Returns the last byte.

Panics

If the Bytes is empty.

pub fn copy_into_slice(&self, slice: &mut [u8; N])

Copy the bytes into the given slice.

pub fn to_array(&self) -> [u8; N]

Copy the bytes in BytesN into an array.

pub fn iter(&self) -> BytesIter