soroban_sdk
Struct Env

pub struct Env { /* private fields */ }

The Env type provides access to the environment the contract is executing within.

The Env provides access to information about the currently executing contract, who invoked it, contract data, functions for signing, hashing, etc.

Most types require access to an Env to be constructed or converted.

Implementations

impl Env

pub fn storage(&self) -> Storage

Get a Storage for accessing and updating persistent data owned by the currently executing contract.

pub fn events(&self) -> Events

Get Events for publishing events associated with the currently executing contract.

pub fn ledger(&self) -> Ledger

Get a Ledger for accessing the current ledger.

pub fn deployer(&self) -> Deployer

Get a deployer for deploying contracts.

pub fn crypto(&self) -> Crypto

Get a Crypto for accessing the current cryptographic functions.

pub fn prng(&self) -> Prng

Get a Prng for accessing the current functions which provide pseudo-randomness.

Warning

The pseudo-random generator returned is not suitable for security-sensitive work.

pub fn current_contract_address(&self) -> Address

Get the Address object corresponding to the current executing contract.

pub fn invoke_contract<T>( &self, contract_address: &Address, func: &Symbol, args: Vec<Val>, ) -> T
where T: TryFromVal<Env, Val>,

Invokes a function of a contract that is registered in the Env.

Panics

Will panic if the contract_id does not match a registered contract, func does not match a function of the referenced contract, or the number of args do not match the argument count of the referenced contract function.

Will panic if the contract that is invoked fails or aborts in anyway.

Will panic if the value returned from the contract cannot be converted into the type T.

pub fn try_invoke_contract<T, E>( &self, contract_address: &Address, func: &Symbol, args: Vec<Val>, ) -> Result<Result<T, T::Error>, Result<E, InvokeError>>
where T: TryFromVal<Env, Val>, E: TryFrom<Error>, E::Error: Into<InvokeError>,

Invokes a function of a contract that is registered in the Env, returns an error if the invocation fails for any reason.

pub fn authorize_as_current_contract( &self, auth_entries: Vec<InvokerContractAuthEntry>, )

Authorizes sub-contract calls on behalf of the current contract.

All the direct calls that the current contract performs are always considered to have been authorized. This is only needed to authorize deeper calls that originate from the next contract call from the current contract.

For example, if the contract A calls contract B, contract B calls contract C and contract C calls A.require_auth(), then an entry corresponding to C call has to be passed in auth_entries. It doesn’t matter if contract B called require_auth or not. If contract A calls contract B again, then authorize_as_current_contract has to be called again with the respective entries.

pub fn logs(&self) -> Logs

Get the Logs for logging debug events.