soroban_sdk
Struct Timepoint

pub struct Timepoint { /* private fields */ }

Timepoint holds a 64-bit unsigned integer.

Implementations

impl Timepoint

pub fn env(&self) -> &Env

pub fn as_val(&self) -> &Val

pub fn to_val(&self) -> Val

pub fn to_val_type(&self) -> TimepointVal

impl Timepoint

pub fn from_unix(env: &Env, seconds: u64) -> Timepoint

Create a Timepoint from a unix time in seconds, the time in seconds since January 1, 1970 UTC.

pub fn to_unix(&self) -> u64

Returns the Timepoint as unix time in seconds, the time in seconds since January 1, 1970 UTC.