soroban_sdk
Struct Val

pub struct Val(/* private fields */);

Raw value of the Soroban smart contract platform that types can be converted to and from for storing, or passing between contracts.

Implementations

impl Val

pub const fn can_represent_scval_type(scv_ty: ScValType) -> bool

Some ScVals are not representable as Vals at all, and only exist in the XDR to serve as special storage system key or value payloads managed by the Host.

pub fn can_represent_scval(scv: &ScVal) -> bool

Non-recursively checks whether ScVal can be represented as Val. Since conversions from ScVal are recursive themselves, this should be called at every recursion level during conversion.

pub fn can_represent_scval_recursive(scv: &ScVal) -> bool

Recursively checks whether ScVal can be represented as Val. This should only be used once per top-level ScVal.

pub fn is_good(self) -> bool

We define a “good” Val as one that has one of the allowed tag values, all the defined body-bits for its case set to valid values, and all the undefined body-bits set to zero.

pub const fn get_payload(self) -> u64

pub const fn from_payload(x: u64) -> Val

pub const fn shallow_eq(&self, other: &Val) -> bool

pub const fn get_tag(self) -> Tag

pub const fn is_object(self) -> bool

pub const fn from_void() -> Void

pub const fn from_bool(b: bool) -> Bool

pub const fn is_void(self) -> bool

pub const fn is_true(self) -> bool

pub const fn is_false(self) -> bool

impl Val

pub const I32_ZERO: I32Val = _

pub const I32_MIN: I32Val = _

pub const I32_MAX: I32Val = _

pub const U32_ZERO: U32Val = _

pub const U32_ONE: U32Val = _

pub const U32_MIN: U32Val = _

pub const U32_MAX: U32Val = _

pub const VOID: Void = _

pub const TRUE: Bool = _

pub const FALSE: Bool = _

impl Val

pub const fn from_u32(u: u32) -> U32Val

pub const fn from_i32(i: i32) -> I32Val