soroban_sdk
Struct Vec

pub struct Vec<T> { /* private fields */ }

Vec is a sequential and indexable growable collection type.

Values are stored in the environment and are available to contract through the functions defined on Vec. Values stored in the Vec are transmitted to the environment as Vals, and when retrieved from the Vec are transmitted back and converted from Val back into their type.

The values in a Vec are not guaranteed to be of type T and conversion will fail if they are not. Most functions on Vec have a try_ variation that returns a Result that will be Err if the conversion fails. Functions that are not prefixed with try_ will panic if conversion fails.

There are some cases where this lack of guarantee is important:

When storing a Vec that has been provided externally as a contract function argument, be aware there is no guarantee that all items in the Vec will be of type T. It may be necessary to validate all values, either before storing, or when loading with try_ variation functions.

When accessing and iterating over a Vec that has been provided externally as a contract function input, and the contract needs to be resilient to failure, use the try_ variation functions.

Functions with an _unchecked suffix will panic if called with indexes that are out-of-bounds.

To store u8s and binary data, use Bytes/BytesN instead.

Vec values can be stored as Storage, or in other types like Vec, Map, etc.

Examples

use soroban_sdk::{vec, Env};

let env = Env::default();
let vec = vec![&env, 0, 1, 2, 3];
assert_eq!(vec.len(), 4);

Implementations

impl<T> Vec<T>

pub fn env(&self) -> &Env

pub fn as_val(&self) -> &Val

pub fn to_val(&self) -> Val

pub fn as_object(&self) -> &VecObject

pub fn to_object(&self) -> VecObject

pub fn to_vals(&self) -> Vec<Val>

impl<T> Vec<T>
where T: IntoVal<Env, Val> + TryFromVal<Env, Val>,

pub fn new(env: &Env) -> Vec<T>

Create an empty Vec.

pub fn from_array<const N: usize>(env: &Env, items: [T; N]) -> Vec<T>

Create a Vec from the array of items.

pub fn from_slice(env: &Env, items: &[T]) -> Vec<T>
where T: Clone,

Create a Vec from the slice of items.

pub fn get(&self, i: u32) -> Option<T>

Returns the item at the position or None if out-of-bounds.

Panics

If the value at the position cannot be converted to type T.

pub fn try_get(&self, i: u32) -> Result<Option<T>, T::Error>

Returns the item at the position or None if out-of-bounds.

Errors

If the value at the position cannot be converted to type T.

pub fn get_unchecked(&self, i: u32) -> T

Returns the item at the position.

Panics

If the position is out-of-bounds.

If the value at the position cannot be converted to type T.

pub fn try_get_unchecked(&self, i: u32) -> Result<T, T::Error>

Returns the item at the position.

Errors

If the value at the position cannot be converted to type T.

Panics

If the position is out-of-bounds.

pub fn set(&mut self, i: u32, v: T)

Sets the item at the position with new value.

Panics

If the position is out-of-bounds.

pub fn remove(&mut self, i: u32) -> Option<()>

Removes the item at the position.

Returns None if out-of-bounds.

pub fn remove_unchecked(&mut self, i: u32)

Removes the item at the position.

Panics

If the position is out-of-bounds.

pub fn push_front(&mut self, x: T)

Adds the item to the front.

Increases the length by one, shifts all items up by one, and puts the item in the first position.

pub fn pop_front(&mut self) -> Option<T>

Removes and returns the first item or None if empty.

Panics

If the value at the first position cannot be converted to type T.

pub fn try_pop_front(&mut self) -> Result<Option<T>, T::Error>

Removes and returns the first item or None if empty.

Errors

If the value at the first position cannot be converted to type T.

pub fn pop_front_unchecked(&mut self) -> T

Removes and returns the first item.

Panics

If the vec is empty.

If the value at the first position cannot be converted to type T.

pub fn try_pop_front_unchecked(&mut self) -> Result<T, T::Error>

Removes and returns the first item.

Errors

If the value at the first position cannot be converted to type T.

Panics

If the vec is empty.

pub fn push_back(&mut self, x: T)

Adds the item to the back.

Increases the length by one and puts the item in the last position.

pub fn pop_back(&mut self) -> Option<T>

Removes and returns the last item or None if empty.

Panics

If the value at the last position cannot be converted to type T.

pub fn try_pop_back(&mut self) -> Result<Option<T>, T::Error>

Removes and returns the last item or None if empty.

Errors

If the value at the last position cannot be converted to type T.

pub fn pop_back_unchecked(&mut self) -> T

Removes and returns the last item.

Panics

If the vec is empty.

If the value at the last position cannot be converted to type T.

pub fn try_pop_back_unchecked(&mut self) -> Result<T, T::Error>

Removes and returns the last item.

Errors

If the value at the last position cannot be converted to type T.

Panics

If the vec is empty.

pub fn first(&self) -> Option<T>

Returns the first item or None if empty.

Panics

If the value at the first position cannot be converted to type T.

pub fn try_first(&self) -> Result<Option<T>, T::Error>

Returns the first item or None if empty.

Errors

If the value at the first position cannot be converted to type T.

pub fn first_unchecked(&self) -> T

Returns the first item.

Panics

If the vec is empty.

If the value at the first position cannot be converted to type T.

pub fn try_first_unchecked(&self) -> Result<T, T::Error>

Returns the first item.

Errors

If the value at the first position cannot be converted to type T.

Panics

If the vec is empty.

pub fn last(&self) -> Option<T>

Returns the last item or None if empty.

Panics

If the value at the last position cannot be converted to type T.

pub fn try_last(&self) -> Result<Option<T>, T::Error>

Returns the last item or None if empty.

Errors

If the value at the last position cannot be converted to type T.

pub fn last_unchecked(&self) -> T

Returns the last item.

Panics

If the vec is empty.

If the value at the last position cannot be converted to type T.

pub fn try_last_unchecked(&self) -> Result<T, T::Error>

Returns the last item.

Errors

If the value at the last position cannot be converted to type T.

Panics

If the vec is empty.

pub fn insert(&mut self, i: u32, x: T)

Inserts an item at the position.

Panics

If the position is out-of-bounds.

pub fn append(&mut self, other: &Vec<T>)

Append the items.

pub fn extend_from_array<const N: usize>(&mut self, items: [T; N])

Extend with the items in the array.

pub fn extend_from_slice(&mut self, items: &[T])
where T: Clone,

Extend with the items in the slice.

impl<T> Vec<T>

pub fn slice(&self, r: impl RangeBounds<u32>) -> Self

Returns a subset of the bytes as defined by the start and end bounds of the range.

Panics

If the range is out-of-bounds.

pub fn shuffle(&mut self)

Returns copy of the vec shuffled using the NOT-SECURE PRNG.

In tests, must be called from within a running contract.

Warning

The pseudo-random generator used to perform the shuffle is not suitable for security-sensitive work.

pub fn to_shuffled(&self) -> Self

Returns copy of the vec shuffled using the NOT-SECURE PRNG.

In tests, must be called from within a running contract.

Warning

The pseudo-random generator used to perform the shuffle is not suitable for security-sensitive work.

pub fn is_empty(&self) -> bool

Returns true if the vec is empty and contains no items.

pub fn len(&self) -> u32

Returns the number of items in the vec.

impl<T> Vec<T>
where T: IntoVal<Env, Val>,

pub fn contains(&self, item: impl Borrow<T>) -> bool

Returns true if the Vec contains the item.

pub fn first_index_of(&self, item: impl Borrow<T>) -> Option<u32>

Returns the index of the first occurrence of the item.

If the item cannot be found None is returned.

pub fn last_index_of(&self, item: impl Borrow<T>) -> Option<u32>

Returns the index of the last occurrence of the item.

If the item cannot be found None is returned.

pub fn binary_search(&self, item: impl Borrow<T>) -> Result<u32, u32>

Returns the index of an occurrence of the item in an already sorted Vec, or the index of where the item can be inserted to keep the Vec sorted.

If the item is found, Result::Ok is returned containing the index of the item.

If the item is not found, Result::Err is returned containing the index of where the item could be inserted to retain the sorted ordering.

impl<T> Vec<Vec<T>>
where T: IntoVal<Env, Val> + TryFromVal<Env, Val> + Clone,

pub fn concat(&self) -> Vec<T>

impl<T> Vec<T>
where T: IntoVal<Env, Val> + TryFromVal<Env, Val>,

pub fn iter(&self) -> UnwrappedIter<VecTryIter<T>, T, T::Error>
where T: IntoVal<Env, Val> + TryFromVal<Env, Val> + Clone, T::Error: Debug,

pub fn try_iter(&self) -> VecTryIter<T>
where T: IntoVal<Env, Val> + TryFromVal<Env, Val> + Clone,

pub fn into_try_iter(self) -> VecTryIter<T>
where T: IntoVal<Env, Val> + TryFromVal<Env, Val> + Clone, T::Error: Debug,