soroban_sdk
Trait TryFromVal

pub trait TryFromVal<E, V>: Sized
where
    E: Env,
    V: ?Sized,
{
    type Error: Debug + Into<Error>;

    // Required method
    fn try_from_val(env: &E, v: &V) -> Result<Self, Self::Error>;
}

Used to do conversions between values in the Soroban environment. Trait for types that can be fallibly converted to another type V, analogous to the standard Rust type TryFrom, but making use of the provided Env implementation E in order to convert parts of the type that require it. Mainly this exists because Val types that contain object handles need to delegate to the environment to look up and extract the content of those handles.

Required Associated Types

type Error: Debug + Into<Error>

Required Methods

fn try_from_val(env: &E, v: &V) -> Result<Self, Self::Error>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.