soroban_sdk

Trait TryIntoVal

pub trait TryIntoVal<E, V>
where
    E: Env,
{
    type Error: Debug + Into<Error>;

    // Required method
    fn try_into_val(&self, env: &E) -> Result<V, Self::Error>;
}

Used to do conversions between values in the Soroban environment. The opposite trait to TryFromVal, analogous to the way that TryInto exists as an opposite to TryFrom. Exists only for convenience of doing conversions via .try_into_val(e) or specifying convertability with a bound like TryIntoVal<E,Other>.

Required Associated Types

type Error: Debug + Into<Error>

Required Methods

fn try_into_val(&self, env: &E) -> Result<V, Self::Error>