soroban_sdk::xdr
Enum Error

pub enum Error {
    Invalid,
    Unsupported,
    LengthExceedsMax,
    LengthMismatch,
    NonZeroPadding,
    Utf8Error(Utf8Error),
    InvalidHex,
    Io(Error),
    DepthLimitExceeded,
    LengthLimitExceeded,
}

Error contains all errors returned by functions in this crate. It can be compared via PartialEq, however any contained IO errors will only be compared on their ErrorKind.

Variants

Invalid

Unsupported

LengthExceedsMax

LengthMismatch

NonZeroPadding

Utf8Error(Utf8Error)

InvalidHex

Io(Error)

DepthLimitExceeded

LengthLimitExceeded