soroban_sdk::xdr
Trait FromXdr

pub trait FromXdr: Sized {
    type Error;

    // Required method
    fn from_xdr(env: &Env, b: &Bytes) -> Result<Self, Self::Error>;
}

Implemented by types that can be deserialized from Bytes.

All types that are convertible from Val are implemented.

Required Associated Types

type Error

Required Methods

fn from_xdr(env: &Env, b: &Bytes) -> Result<Self, Self::Error>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.